fix: address yarr rebrand review findings
