+++ CHANGELOG.md
@@ -500,7 +500,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `.github/workflows/ci.yml` — CI: fmt, clippy, nextest, taplo, audit, gitleaks
 - `.github/workflows/docker-publish.yml` — multi-platform Docker build + Trivy scan
 - `.github/workflows/release.yml` — release binaries for linux/amd64 and linux/arm64
-- `config.rustarr.toml` — fully annotated config template
+- `config.yarr.toml` — fully annotated config template
 - `.env.rustarr` — documented secrets template
 - `CHANGELOG.md` following Keep a Changelog format
 - Workspace structure: root crate + `xtask/` member
