+++ xtask/src/live/contract.rs
@@ -10,6 +10,8 @@
 //! Operations run GET -> POST -> PUT/PATCH -> DELETE so reads/updates see existing
 //! resources before deletes remove them. Pass `--no-destructive` to skip DELETEs.
 
+pub(super) mod invoke;
+pub(super) mod seeding;
 pub(super) mod synth;
 
 use anyhow::{Context, Result};
@@ -21,6 +23,9 @@ use yarr::ServiceKind;
 use yarr::openapi::{self, HttpMethod, OperationSpec};
 
 use super::{process, report};
+pub(super) use invoke::is_retryable_contract_error;
+use invoke::{invoke, write_detail};
+pub(super) use seeding::seed_service_fixtures;
 use synth::Spec;
 
 /// (kind str, spec path) for the spec-backed services.
@@ -126,195 +131,6 @@ pub fn run(
             report.fail(format!("contract {svc}"), status.detail);
         }
     }
-    Ok(())
-}
-
-pub(super) fn seed_service_fixtures(
-    yarr: &process::YarrProcess,
-    svc: &str,
-    kind: ServiceKind,
-) -> Result<()> {
-    match kind {
-        ServiceKind::Sonarr => ensure_sonarr_download_client(yarr, svc),
-        _ => Ok(()),
-    }
-}
-
-fn ensure_sonarr_download_client(yarr: &process::YarrProcess, svc: &str) -> Result<()> {
-    ensure_sonarr_qbittorrent_download_client(yarr, svc)?;
-    ensure_sonarr_newznab_indexer(yarr, svc)?;
-    ensure_sonarr_custom_script_notification(yarr, svc)?;
-    ensure_sonarr_remote_path_mapping(yarr, svc)?;
-    ensure_sonarr_autotagging(yarr, svc)
-}
-
-fn ensure_sonarr_qbittorrent_download_client(yarr: &process::YarrProcess, svc: &str) -> Result<()> {
-    let existing = yarr.json(&[svc, "op", "get_downloadclient", "--args", "{}"])?;
-    if existing.as_array().is_some_and(|items| {
-        items.iter().any(|item| {
-            item.get("name").and_then(Value::as_str) == Some("yarr-live-qbit")
-                && item.get("implementation").and_then(Value::as_str) == Some("QBittorrent")
-        })
-    }) {
-        return Ok(());
-    }
-    let body = json!({
-        "enable": false,
-        "protocol": "torrent",
-        "priority": 1,
-        "removeCompletedDownloads": false,
-        "removeFailedDownloads": false,
-        "name": "yarr-live-qbit",
-        "implementation": "QBittorrent",
-        "implementationName": "qBittorrent",
-        "configContract": "QBittorrentSettings",
-        "fields": [
-            {"name": "host", "value": "100.118.209.1"},
-            {"name": "port", "value": 8080},
-            {"name": "useSsl", "value": false},
-            {"name": "urlBase", "value": ""},
-            {"name": "apiKey", "value": ""},
-            {"name": "username", "value": ""},
-            {"name": "password", "value": ""},
-            {"name": "tvCategory", "value": "tv-sonarr"},
-            {"name": "tvImportedCategory", "value": ""},
-            {"name": "recentTvPriority", "value": 0},
-            {"name": "olderTvPriority", "value": 0},
-            {"name": "initialState", "value": 0},
-            {"name": "sequentialOrder", "value": false},
-            {"name": "firstAndLast", "value": false},
-            {"name": "contentLayout", "value": 0}
-        ],
-        "tags": []
-    });
-    let args = serde_json::to_string(&json!({ "body": body }))?;
-    yarr.json(&[svc, "op", "post_downloadclient", "--args", &args])?;
-    Ok(())
-}
-
-fn ensure_sonarr_newznab_indexer(yarr: &process::YarrProcess, svc: &str) -> Result<()> {
-    let existing = yarr.json(&[svc, "op", "get_indexer", "--args", "{}"])?;
-    if existing.as_array().is_some_and(|items| {
-        items.iter().any(|item| {
-            item.get("name").and_then(Value::as_str) == Some("yarr-live-newznab")
-                && item.get("implementation").and_then(Value::as_str) == Some("Newznab")
-        })
-    }) {
-        return Ok(());
-    }
-    let body = json!({
-        "enableRss": false,
-        "enableAutomaticSearch": false,
-        "enableInteractiveSearch": false,
-        "supportsRss": true,
-        "supportsSearch": true,
-        "protocol": "usenet",
-        "priority": 1,
-        "name": "yarr-live-newznab",
-        "implementation": "Newznab",
-        "implementationName": "Newznab",
-        "configContract": "NewznabSettings",
-        "fields": [
-            {"name": "baseUrl", "value": "http://127.0.0.1:9"},
-            {"name": "apiPath", "value": "/api"},
-            {"name": "apiKey", "value": "yarr-live"},
-            {"name": "categories", "value": [5030, 5040]},
-            {"name": "animeCategories", "value": []},
-            {"name": "animeStandardFormatSearch", "value": false},
-            {"name": "additionalParameters", "value": ""},
-            {"name": "multiLanguages", "value": []},
-            {"name": "failDownloads", "value": []}
-        ],
-        "tags": []
-    });
-    let args = serde_json::to_string(&json!({ "body": body }))?;
-    yarr.json(&[svc, "op", "post_indexer", "--args", &args])?;
-    Ok(())
-}
-
-fn ensure_sonarr_custom_script_notification(yarr: &process::YarrProcess, svc: &str) -> Result<()> {
-    let existing = yarr.json(&[svc, "op", "get_notification", "--args", "{}"])?;
-    if existing.as_array().is_some_and(|items| {
-        items.iter().any(|item| {
-            item.get("name").and_then(Value::as_str) == Some("yarr-live-script")
-                && item.get("implementation").and_then(Value::as_str) == Some("CustomScript")
-        })
-    }) {
-        return Ok(());
-    }
-    let body = json!({
-        "name": "yarr-live-script",
-        "implementation": "CustomScript",
-        "implementationName": "Custom Script",
-        "configContract": "CustomScriptSettings",
-        "fields": [
-            {"name": "path", "value": "/bin/true"},
-            {"name": "arguments", "value": ""}
-        ],
-        "onGrab": false,
-        "onDownload": false,
-        "onUpgrade": false,
-        "onRename": false,
-        "onSeriesAdd": false,
-        "onSeriesDelete": false,
-        "onEpisodeFileDelete": false,
-        "onEpisodeFileDeleteForUpgrade": false,
-        "onHealthIssue": false,
-        "onHealthRestored": false,
-        "onApplicationUpdate": false,
-        "includeHealthWarnings": false,
-        "tags": []
-    });
-    let args = serde_json::to_string(&json!({ "body": body }))?;
-    yarr.json(&[svc, "op", "post_notification", "--args", &args])?;
-    Ok(())
-}
-
-fn ensure_sonarr_remote_path_mapping(yarr: &process::YarrProcess, svc: &str) -> Result<()> {
-    let existing = yarr.json(&[svc, "op", "get_remotepathmapping", "--args", "{}"])?;
-    if existing.as_array().is_some_and(|items| {
-        items.iter().any(|item| {
-            item.get("host").and_then(Value::as_str) == Some("yarr-live-host")
-                && item.get("remotePath").and_then(Value::as_str) == Some("/downloads/")
-                && item.get("localPath").and_then(Value::as_str) == Some("/data/media/tv/")
-        })
-    }) {
-        return Ok(());
-    }
-    let body = json!({
-        "host": "yarr-live-host",
-        "remotePath": "/downloads/",
-        "localPath": "/data/media/tv/"
-    });
-    let args = serde_json::to_string(&json!({ "body": body }))?;
-    yarr.json(&[svc, "op", "post_remotepathmapping", "--args", &args])?;
-    Ok(())
-}
-
-fn ensure_sonarr_autotagging(yarr: &process::YarrProcess, svc: &str) -> Result<()> {
-    let existing = yarr.json(&[svc, "op", "get_autotagging", "--args", "{}"])?;
-    if existing.as_array().is_some_and(|items| {
-        items
-            .iter()
-            .any(|item| item.get("name").and_then(Value::as_str) == Some("yarr-live-autotag"))
-    }) {
-        return Ok(());
-    }
-    let body = json!({
-        "name": "yarr-live-autotag",
-        "removeTagsAutomatically": false,
-        "tags": [40],
-        "specifications": [{
-            "name": "Monitored",
-            "implementation": "MonitoredSpecification",
-            "implementationName": "Monitored",
-            "negate": false,
-            "required": false,
-            "fields": []
-        }]
-    });
-    let args = serde_json::to_string(&json!({ "body": body }))?;
-    yarr.json(&[svc, "op", "post_autotagging", "--args", &args])?;
     Ok(())
 }
 
@@ -937,104 +753,6 @@ fn is_unseeded_optional_feature_endpoint(kind: ServiceKind, path: &str) -> bool 
         }
         _ => false,
     }
-}
-
-/// Invoke `yarr <svc> op <name> --args <json> [--confirm]`. Returns the parsed
-/// JSON result on a 2xx, `None` for an empty body, or an error with the upstream
-/// message on a non-2xx / CLI error.
-const CONTRACT_INVOKE_ATTEMPTS: usize = 3;
-
-fn invoke(
-    yarr: &process::YarrProcess,
-    svc: &str,
-    name: &str,
-    args: &Map<String, Value>,
-    confirm: bool,
-) -> Result<Option<Value>> {
-    let mut last_error = None;
-    for attempt in 1..=CONTRACT_INVOKE_ATTEMPTS {
-        match invoke_once(yarr, svc, name, args, confirm) {
-            Ok(value) => return Ok(value),
-            Err(err) => {
-                let detail = err.to_string();
-                if attempt < CONTRACT_INVOKE_ATTEMPTS && is_retryable_contract_error(&detail) {
-                    last_error = Some(detail);
-                    std::thread::sleep(std::time::Duration::from_millis(750));
-                    continue;
-                }
-                if attempt > 1 {
-                    let prior = last_error
-                        .map(|e| format!("; previous retryable error: {e}"))
-                        .unwrap_or_default();
-                    anyhow::bail!("after {attempt} attempts: {detail}{prior}");
-                }
-                return Err(err);
-            }
-        }
-    }
-    unreachable!("contract invoke loop always returns");
-}
-
-pub(super) fn is_retryable_contract_error(detail: &str) -> bool {
-    detail.contains("request failed")
-        || detail.contains("tcp connect error")
-        || detail.contains("connection closed")
-        || detail.contains("error sending request")
-}
-
-fn invoke_once(
-    yarr: &process::YarrProcess,
-    svc: &str,
-    name: &str,
-    args: &Map<String, Value>,
-    confirm: bool,
-) -> Result<Option<Value>> {
-    let args_json = serde_json::to_string(args)?;
-    let mut argv: Vec<&str> = vec![svc, "op", name, "--args", &args_json];
-    if confirm {
-        argv.push("--confirm");
-    }
-    let output = yarr.output(&argv)?;
-    if !output.status.success() {
-        let err = String::from_utf8_lossy(&output.stderr);
-        anyhow::bail!("{}", err.trim().trim_start_matches("Error: "));
-    }
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let trimmed = stdout.trim();
-    if trimmed.is_empty() {
-        return Ok(None);
-    }
-    // A non-empty 2xx body MUST parse as JSON. Swallowing a parse error here (the old
-    // `.ok()`) made unparseable output masquerade as an empty body, silently SKIPPING
-    // schema validation and counting as a clean pass — a false PASS. Surface it as a
-    // failure with a preview of the offending output instead.
-    let value: Option<Value> = match serde_json::from_str(trimmed) {
-        Ok(v) => Some(v),
-        Err(e) => anyhow::bail!(
-            "non-empty 2xx body did not parse as JSON ({e}): {}",
-            trimmed.chars().take(180).collect::<String>()
-        ),
-    };
-    // `YarrClient` returns `{"ok":true,"status":<code>}` for an empty 2xx body
-    // (204 etc.). That's a "no body" sentinel, not a response to validate against
-    // the op's schema — treat it like an empty body so it counts as a clean 2xx.
-    if let Some(Value::Object(m)) = &value
-        && m.len() == 2
-        && m.get("ok") == Some(&Value::Bool(true))
-        && m.get("status").is_some_and(Value::is_number)
-    {
-        return Ok(None);
-    }
-    Ok(value)
-}
-
-fn write_detail(svc: &str, results: &[OpResult]) -> Result<()> {
-    let dir = std::path::Path::new("target/live-full");
-    std::fs::create_dir_all(dir)?;
-    let path = dir.join(format!("contract-{svc}.json"));
-    std::fs::write(&path, serde_json::to_string_pretty(results)?)
-        .with_context(|| format!("write {}", path.display()))?;
-    Ok(())
 }
 
 #[cfg(test)]
